# distributed-system-v2/repo-stat/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Distributed System API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:28080",
    "basePath": "/",
    "paths": {
        "/api/ping": {
            "get": {
                "description": "Check status of processor and subscriber",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "ping services status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PingResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.PingResponse"
                        }
                    }
                }
            }
        },
        "/api/repositories/info": {
            "get": {
                "description": "Get GitHub repository information by URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repositories"
                ],
                "summary": "Get repository info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub repository URL",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.RepositoryInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Owner": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                }
            }
        },
        "domain.RepositoryInfo": {
            "type": "object",
            "properties": {
                "creationDate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "forksCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/domain.Owner"
                },
                "stargazersCount": {
                    "type": "integer"
                }
            }
        },
        "dto.PingResponse": {
            "type": "object",
            "properties": {
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ServiceDTO"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.ServiceDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}